adjusted stories coz I broke it with the refactor
