+++ frontend/src/stories/CardDetails.ProjectDetails.stories.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import ProjectDetails from "../components/regions/CardDetails/ProjectDetails";
+import ProjectDetails from "../components/pages/CardDetails/ProjectDetails";
 import reviewObject from "./fixtures/review.json";
 import projectObject from "./fixtures/project.json";
 import { Provider } from "react-redux";
