+++ frontend/src/stories/CardDetails.Review.stories.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Review from "../components/regions/CardDetails/Review";
+import Review from "../components/pages/CardDetails/Review";
 import review from "./fixtures/review.json";
 
 export default {
