+++ frontend/src/stories/CardDetails.Reviews.stories.js
@@ -1,6 +1,6 @@
 import React from "react";
 
-import Reviews from "../components/regions/CardDetails/Reviews";
+import Reviews from "../components/pages/CardDetails/Reviews";
 
 import reviewObject from "./fixtures/review.json";
